CMakeLists.txt: use "full RPATH" settings
This commit adds some magic to the toplevel CMake file to configure
the shared libraries and executables to have RPATH information
inserted into them upon installation. This makes it so that the core
shared libraries can be located by things that depend upon them
without needing to muck with LD_LIBRARY_PATH and friends, if the
libraries are installed in a nonstandard prefix.

The commands come from: http://www.cmake.org/Wiki/CMake_RPATH_handling